djangotest/requirements.txt
Django==2.2.11
sendgrid==5.6.0
braintree==4.3.0
djangorestframework==3.11.1
django-cors-headers==3.4.0
